change label list to int for clusterless thresholder spike sorter

+++ src/spyglass/spikesorting/spikesorting_sorting.py
@@ -176,7 +176,7 @@ class SpikeSorting(dj.Computed):
             sorter_params.pop('tempdir',None)
             detected_spikes = detect_peaks(recording, **sorter_params)
             sorting = si.NumpySorting.from_times_labels(times_list=detected_spikes['sample_ind'],
-                                                        labels_list=np.zeros(len(detected_spikes)),
+                                                        labels_list=np.zeros(len(detected_spikes), dtype=np.int),
                                                         sampling_frequency=recording.get_sampling_frequency())
         else:
             sorting = sis.run_sorter(sorter, recording,
